Added support for importing vault exported content.
Modified ProsperSpec so the setupSpec method will import vault content if it exists.  The import will look for a test/resources/test-content directory and if it exists will import the vault content structure that exists below it.  The test-content directory must contain a child jcr_root directory and META-INF directory.  The jcr_root directory will contain all the vault exported content.  The only file required to be in the META-INF directory is the vault/filter.xml file.  A future enhancement will look into providing a way to dynamically generate the filter rules so that each test can target a subset of content from the test-content directory.